+++ src/Pim/Bundle/CatalogBundle/Doctrine/Query/ProductQueryBuilderInterface.php
@@ -2,8 +2,6 @@
 
 namespace Pim\Bundle\CatalogBundle\Doctrine\Query;
 
-use Pim\Bundle\CatalogBundle\Model\AbstractAttribute;
-
 /**
  * Aims to customize a query builder to add useful shortcuts which allow to easily select, filter or sort a product
  * values
@@ -28,22 +26,14 @@ interface ProductQueryBuilderInterface
     public function addFilter($field, $operator, $value);
 
     /**
-     * Sort by attribute value
-     *
-     * @param AbstractAttribute $attribute the attribute to sort on
-     * @param string            $direction the direction to use
-     *
-     * @return ProductQueryBuilderInterface
-     */
-    public function addAttributeSorter(AbstractAttribute $attribute, $direction);
-
-    /**
      * Sort by field
      *
      * @param string $field     the field to sort on
      * @param string $direction the direction to use
      *
+     * @throws \LogicException
+     *
      * @return ProductQueryBuilderInterface
      */
-    public function addFieldSorter($field, $direction);
+    public function addSorter($field, $direction);
 }